Add Loader and Linker for types.
